feat: move some components to ui package
