--- apps/web/components/atoms/link/link.stories.tsx
@@ -1,16 +1,0 @@
-import { Link } from 'atoms/link/link';
-import { ComponentMeta, ComponentStory } from '@storybook/react';
-
-export default {
-	title: 'Main Library/atoms/Link',
-	component: Link,
-} as ComponentMeta<typeof Link>;
-
-export const Default: ComponentStory<typeof Link> = (props) => {
-	return <Link {...props}>Zaloguj się</Link>;
-};
-
-Default.args = {
-	href: '',
-	variant: 'primary',
-};
