+++ apps/web/components/molecules/fileInput/fileInput.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, ReactNode } from 'react';
 import { useFileInput } from 'molecules/fileInput/useFileInput';
-import { Text } from 'atoms/text/text';
+import { Text } from 'ui';
 
 type FileInputProps = {
 	handleChange: (filesSrc: string[]) => void;
