+++ packages/ui/components/atoms/text/text.stories.tsx
@@ -6,6 +6,22 @@ export default {
 	component: Text,
 } as ComponentMeta<typeof Text>;
 
-export const Default: ComponentStory<typeof Text> = (props) => {
-	return <Text {...props}>Example message</Text>;
+export const Primary: ComponentStory<typeof Text> = (props) => {
+	return <Text {...props}>Example text</Text>;
+};
+
+Primary.args = {
+	variant: 'default',
+	size: 'medium',
+	tag: 'p',
+};
+
+export const ErrorText: ComponentStory<typeof Text> = (props) => {
+	return <Text {...props}>Example text</Text>;
+};
+
+ErrorText.args = {
+	variant: 'error',
+	size: 'medium',
+	tag: 'p',
 };
